AB#92103 Add possibility to use api key in query string
For some clients of WFS and MVT api's it can be hard or even
impossible to use api keys using http headers.

It is now also possible to pass in the api key as a querystring
with the name `x-api-key`. The same JWT token can be used.
The JWT tokens are url-safe. A possible caveat could be that
the url gets too long, however most browsers/tools are supporting
fairly largs urls (about 2K chars).

The api key middleware now also does a (deliberately) simple user agent check
on incoming requests. Otherwise the API frontpage + docs are not
accessible without an API key.