AB#91451 Handle situation that apikeyserver is down
If keys are mandatory, we must have the apikeyserver running,
because we need to be able to fetch the signing keys for checking
the incoming api keys.

If api keys are not mandatory, we do not bail out, however
if no signing keys are available, we cannot do any checks.
So, we add an extra warning on every api key check
to make it as clear as possible in the logs,
that no checking can be done.