AB#91451 Do not bail out if connection fails from thread
When the api key middleware is initialized and the keys are mandatory,
an exception is raised if there is no connection to the api key server.

However, when the dso api is already running, and the api key server
is going down, the keys are already connected, so dso can keep running.

When the middleware tries to refresh the keys in this situation,
and this is not possible, because the key server is down,
an exception should *not* be raised, because this would silently
kill the thread and no signing key refreshing is possible anymore,
until the DSO gets restarted.

Maybe this situation is a bit hypothetical, because we run DSO
with uwsgi and UWSGI_HARAKIRI has been set to a relatively short
period of 1800 seconds. But, better safe than sorry.